requirements_sqlite.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
alembic==1.12.1
# Frontend dependencies (served via CDN)
# Chart.js is included via CDN in the template